pyproject.toml
[project]
name = "nvsx"
version = "0.1.0"
description = "NVSentinel eXtensions — cinematic runbooks for GPU cluster fault remediation"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "Nav" }]
keywords = ["nvidia", "nvsentinel", "kubernetes", "gpu", "runbook", "operator", "cli", "dcgm"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: System Administrators",
  "Operating System :: POSIX",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: System :: Monitoring",
  "Topic :: System :: Systems Administration",
]
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pydantic>=2.5",
  "pyyaml>=6.0",
  "pymongo>=4.6",
]

[project.optional-dependencies]
convert = ["anthropic>=0.40"]

[project.scripts]
nvsx = "nvsx.cli:main"

[project.urls]
Homepage = "https://github.com/clockwork-io/nvsx"
Issues = "https://github.com/clockwork-io/nvsx/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nvsx"]

[tool.hatch.build.targets.sdist]
include = [
  "src/",
  "runbooks/",
  "shims/",
  "scripts/",
  "examples/",
  "demo/",
  "nvsx",
  "LICENSE",
  "README.md",
]
